+++ markymark/conf.py
@@ -8,18 +8,6 @@ MARKYMARK_EXTENSIONS = getattr(
     defaults.DEFAULT_MARKYMARK_EXTENSIONS
 )
 
-MARKYMARK_CSS = getattr(
-    settings,
-    'MARKYMARK_CSS',
-    defaults.DEFAULT_MARKYMARK_CSS
-)
-
-MARKYMARK_JS = getattr(
-    settings,
-    'MARKYMARK_JS',
-    defaults.DEFAULT_MARKYMARK_JS
-)
-
 MARKYMARK_TEMPLATES = getattr(
     settings,
     'MARKYMARK_TEMPLATES',