+++ markymark/defaults.py
@@ -3,20 +3,6 @@ DEFAULT_MARKYMARK_EXTENSIONS = [
 ]
 
 
-DEFAULT_MARKYMARK_CSS = [
-    'markdown/css/markdown-editor.css',
-    'markdown/css/markdown-editor-adminfix.css',
-]
-
-
-DEFAULT_MARKYMARK_JS = [
-    'markdown/js/markdown-init.js',
-    'markdown/js/markdown.js',
-    'markdown/js/markdown-editor.js',
-    'markdown/js/plugins/clean.js'
-]
-
-
 DEFAULT_MARKYMARK_TEMPLATES = {
     'filer': 'markdown/file.html',
     'anylink': 'markdown/anylink.html',