+++ markymark/extensions/clean.py
@@ -1,0 +1,10 @@
+from markymark.extensions import MarkymarkExtension
+
+
+class CleanExtension(MarkymarkExtension):
+    class Media:
+        js = ('markdown/js/plugins/clean.js',)
+
+
+def makeExtension(**kwargs):
+    return CleanExtension(**kwargs)