+++ markymark/tests/settings.py
@@ -45,19 +45,5 @@ MARKYMARK_EXTENSIONS = [
     'markymark.extensions.contrib.anylink',
     'markymark.extensions.contrib.filer',
     'markymark.extensions.autolink',
-]
-
-MARKYMARK_CSS = [
-    'markdown/css/markdown-editor.css',
-    'markdown/css/markdown-editor-adminfix.css',
-    'markdown/css/plugins/filer-file.css',
-]
-
-MARKYMARK_JS = [
-    'markdown/js/markdown-init.js',
-    'markdown/js/markdown.js',
-    'markdown/js/markdown-editor.js',
-    'markdown/js/plugins/clean.js',
-    'markdown/js/plugins/anylink-link.js',
-    'markdown/js/plugins/filer-file.js',
+    'markymark.extensions.clean',
 ]